refactor(services): extract web api provider clients
Twenty-one methods covering API agent registration, OnePiece provider config and
profiles, endpoint metadata, local endpoint discovery and Hybrid Routing move out
of the composition root, split across four modules so none exceeds 300 lines.

The four `let` bindings they share (`webOnePieceProviderConfig`,
`webOnePieceProviderProfiles`, `nextOnePieceProviderProfileId`,
`webHybridRoutingRules`) live in `web-api-provider-state.ts` and stay unexported;
the modules reach them through accessors. That is what lets one bounded context
span four files without any of them holding a second copy of the mock world.

`deleteApiAgent` stays inline. It cascades across sessions, memories, Loop
definitions and skill bindings, so by the change's own cut rule it belongs to no
single group; it now drops the provider config through an exported behavior
(`deleteWebApiAgentProviderConfig`) rather than reaching for the container.

`slugify` moves to `web-mock-identifiers.ts` because the composition root and the
extracted registration path both need it and neither may import from the other.

Aggregate 18,195 -> 18,313: seven new files' import blocks, four narrow
interfaces' own type imports, and the state accessors. Method bodies are moved
verbatim.
