+++ src/services/web-mock-identifiers.ts
@@ -1,0 +1,6 @@
+export function slugify(value: string) {
+  return value
+    .toLowerCase()
+    .replace(/[^a-z0-9]+/g, "-")
+    .replace(/^-+|-+$/g, "");
+}
